Hope I remember all of the markdown syntax.
Added 2 vite plugins, dts and file copy.  Created vite config and loader to build the element.  Added script command to build the button element.

I thought it would be **harder** to have it build all the elements but only make one of them actually usable as a custom element, but this should give you the basic idea of the steps necessary to create a custom element using a vite.config with a ts file that loads the element.

To build:

  - git pull
  - npm install
  - npm run build:button

There are currently no build steps to automatically create an npm package and upload it for distribution

To create the npm package:
  - cd dist/button
  - npm pack

To use the npm package with a framework (like vue or angular):
  - copy the dist/button/uids-4.0.0-alpha1.tgz.tgz file created by npm pack into your new app
  - npm install --save uids-4.0.0-alpha1.tgz

The js files should now be in a uids folder in your node_modules folder.

To use the custom elements in your framework based app, in your imports do something like:
````
  import HelloWorld from './components/HelloWorld.vue';
  import "/node_modules/uids/banner/banner.umd.cjs";
  import "/node_modules/uids/button/button.umd.cjs";
````

To use the custom elements in a generic html file, upload the files to your server and in the `<head></head>` section of your html, do something like:
````
  <script type="module" src="./button/button.umd.js"></script>
  <script nomodule src="./button/button.js"></script>
````

**N.B.**  The server I am using does not support cjs files in its mime-types so I had to rename the file.  It is probably better to create both *.umd.cjs and *.umd.js files in the vite config so that people using plain html have fewer steps to implement the elements, but for an MVP renaming the file is fine.